[#5] test: 지갑생성, 지갑조회, 머니충전 스프링부트 테스트로 전환
