+++ money-service/src/main/java/com/hello/money/v1/dto/AddMoneyRequest.java
@@ -1,0 +1,11 @@
+package com.hello.money.v1.dto;
+
+import org.springframework.util.Assert;
+
+import java.math.BigInteger;
+
+public record AddMoneyRequest(BigInteger amount) {
+  public AddMoneyRequest {
+    Assert.isTrue(amount.compareTo(BigInteger.ZERO) > 0, "금액은 0보다 커야 합니다.");
+  }
+}
