+++ money-service/src/main/java/com/hello/money/v1/service/WalletPort.java
@@ -1,0 +1,10 @@
+package com.hello.money.v1.service;
+
+import com.hello.money.domain.Wallet;
+
+public interface WalletPort {
+
+  Wallet saveWallet(Wallet wallet);
+
+  Wallet findWalletByAccountId(Long accountId);
+}
